refactor(storage): relocate EngineSnapshot wire DTOs to me-types (JC#3, S2 2b)
EngineSnapshot/EmbeddingSpaceSnapshot/FactVectorSnapshot were left in the
facade's inspect/types.rs by the S1 slice, tagged "stay in the monolith" —
but the me-backend-sqlite carve (this sub-PR) moves dump.rs's
stream_snapshot/dump_json production code below the seam, and it
constructs these types directly. A backend crate cannot reach back into
the facade for them, so they join the other already-relocated inspect
DTOs (DumpFormat/EngineStatistics/the *Stats tree) in me-types.

The facade's inspect/types.rs re-exports all of them unchanged, so
crate::inspect::types::* and every public memory_engine:: path (CLI/MCP
import-export included) keep resolving with no consumer-visible change.
Confirmed via maintainer sign-off (JC#3) before touching inspect/types.rs,
which the carve spec otherwise lists as unchanged.
