+++ memory-engine/lib/memory-engine/src/inspect/types.rs
@@ -3,18 +3,18 @@ use std::collections::BTreeMap;
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 
-use crate::types::{
-    Edge, EmbeddingFingerprint, Event, EventType, Fact, LineageSnapshotEntry, ScopeNode, Summary,
-};
+use crate::types::{Edge, Event, EventType, Fact, LineageSnapshotEntry, ScopeNode, Summary};
 
 /// `DumpFormat`, `EngineStatistics`, and the `*Stats` tree moved to `me-types` (Wave 2
-/// #816 E.4b Phase B) as pure serde DTOs; re-exported here so `crate::inspect::types::*`
-/// and (via `inspect::mod`'s `pub use types::*`) `crate::inspect::*` keep resolving.
-/// `EngineSnapshot`/`EmbeddingSpaceSnapshot`/`FactVectorSnapshot` stay in this module —
-/// they are import/export wire types tied to the engine's own snapshot machinery.
+/// #816 E.4b Phase B) as pure serde DTOs; `EngineSnapshot`/`EmbeddingSpaceSnapshot`/
+/// `FactVectorSnapshot` joined them in the S2 me-backend-sqlite carve (sub-PR 2b) — the
+/// backend's `stream_snapshot`/`dump_json` construct them directly and cannot reach back
+/// into the facade for them. All re-exported here so `crate::inspect::types::*` and (via
+/// `inspect::mod`'s `pub use types::*`) `crate::inspect::*` keep resolving for every
+/// consumer (CLI/MCP import-export included).
 pub use me_types::types::inspect::{
-    DumpFormat, EdgeStats, EngineStatistics, EventStats, FactStats, ScopeStats, StorageStats,
-    SummaryStats,
+    DumpFormat, EdgeStats, EmbeddingSpaceSnapshot, EngineSnapshot, EngineStatistics, EventStats,
+    FactStats, FactVectorSnapshot, ScopeStats, StorageStats, SummaryStats,
 };
 
 // ---------------------------------------------------------------------------
@@ -154,80 +154,11 @@ pub enum ReplayOrder {
 }
 
 // ---------------------------------------------------------------------------
-// Dump
+// Dump / Snapshot / Statistics
 // ---------------------------------------------------------------------------
 //
-// `DumpFormat` moved to `me-types` — see the re-export near the top of this file.
-
-/// One embedding-space registry row in a snapshot (#622).
-///
-/// Carries the `embedding_spaces` table content so the embedding identity survives a
-/// dump→restore. Before #622 the identity was a `config` row (`embedding_meta`) that
-/// round-tripped through the generic config copy; it now lives in its own table, so it is
-/// serialized explicitly here. A pre-#622 snapshot has no `embedding_spaces` field — restore
-/// falls back to translating the legacy `embedding_meta` config value.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct EmbeddingSpaceSnapshot {
-    /// Space name (PK); `"default"` for the degenerate single space.
-    pub name: String,
-    /// Lifecycle status TEXT: `active` / `populating` / `deprecated`.
-    pub status: String,
-    /// The canonical identity tuple, flattened to the row's identity columns.
-    #[serde(flatten)]
-    pub fingerprint: EmbeddingFingerprint,
-}
-
-/// One `fact_vectors` row in a snapshot (#623 background reconstruction).
-///
-/// Carries a **non-active** space's per-fact vector: the `populating` space's
-/// vectors mid-reconstruction, or a `deprecated` space's vectors retained for
-/// rollback after a promote. The **active** space's vectors stay in
-/// `facts[].embedding` (the [`EngineSnapshot::facts`] rows are unchanged), so this
-/// section is purely additive — a pre-#623 snapshot has no `fact_vectors` field
-/// and restore defaults it empty.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct FactVectorSnapshot {
-    /// Owning fact id (FK → `facts.id`).
-    pub fact_id: i64,
-    /// Owning space name (FK → `embedding_spaces.name`).
-    pub space_id: String,
-    /// The stored vector.
-    pub embedding: Vec<f32>,
-}
-
-/// Complete snapshot of engine state.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct EngineSnapshot {
-    pub schema_version: u32,
-    pub storage_epoch: u16,
-    pub embed_dim: usize,
-    pub facts: Vec<Fact>,
-    pub edges: Vec<Edge>,
-    pub summaries: Vec<Summary>,
-    pub scopes: Vec<ScopeNode>,
-    pub events: Vec<Event>,
-    /// Lineage records for promoted wisdom facts (Phase 5a).
-    /// Absent in pre-v8 snapshots — defaults to empty.
-    #[serde(default)]
-    pub lineage: Vec<LineageSnapshotEntry>,
-    /// Embedding-space registry rows (#622). Absent in pre-#622 snapshots — defaults to
-    /// empty, and restore then reconstructs the identity from the legacy `embedding_meta`
-    /// config value.
-    #[serde(default)]
-    pub embedding_spaces: Vec<EmbeddingSpaceSnapshot>,
-    /// `fact_vectors` rows for the non-active embedding spaces (#623). Absent in
-    /// pre-#623 snapshots — defaults to empty (the active vectors live in
-    /// `facts[].embedding`, so an old snapshot loses nothing).
-    #[serde(default)]
-    pub fact_vectors: Vec<FactVectorSnapshot>,
-    pub config: BTreeMap<String, String>,
-}
-
-// ---------------------------------------------------------------------------
-// Statistics
-// ---------------------------------------------------------------------------
-//
-// `EngineStatistics` and the `*Stats` tree moved to `me-types` — see the
+// `DumpFormat`, `EngineSnapshot` (+ `EmbeddingSpaceSnapshot`/`FactVectorSnapshot`),
+// `EngineStatistics`, and the `*Stats` tree all moved to `me-types` — see the
 // re-export near the top of this file.
 
 #[cfg(test)]
